src/retrievers/chatbot/requirement.txt
openai==0.28.1
PyPDF2~=2.0
numpy==1.26.0
python-dotenv
python-multipart
httptools
# uvloop
boto3==1.28.17
langchain==0.0.301
requests==2.31.0
trubrics[streamlit]
streamlit==1.27.0
streamlit_authenticator==0.2.3
spacy==3.6.1
tiktoken==0.5.1
mutagen==1.47.0
func_timeout==4.3.5
s3fs==2023.9.2
git+https://github.com/streamlit/files-connection
streamlit_cognito_auth
yt_dlp
pydub
librosa
chromadb
chainlit